Move the shared cluster MSI key vault to its own ARM template
